Cleanup code and refactor
Stuff is still broken, maybe more than before :P
