+++ int-ui/int-ui-standalone/src/main/kotlin/org/jetbrains/jewel/intui/standalone/styling/IntUiComboBoxStyling.kt
@@ -278,26 +278,22 @@ public object IntUiUndecoratedComboBoxColorsFactory {
 }
 
 public fun ComboBoxMetrics.Companion.default(
-    arrowMinSize: DpSize = DpSize(28.dp, 22.dp),
-    arrowSize: DpSize = DpSize(16.dp, 16.dp),
-    minSize: DpSize = DpSize((49 + 23 + 6).dp, 28.dp),
+    arrowAreaSize: DpSize = DpSize(23.dp, 24.dp),
+    minSize: DpSize = DpSize(49.dp, 24.dp),
     cornerSize: CornerSize = CornerSize(4.dp),
     contentPadding: PaddingValues = PaddingValues(horizontal = 6.dp, vertical = 2.dp),
     borderWidth: Dp = 1.dp,
     maxPopupHeight: Dp = 200.dp,
-): ComboBoxMetrics =
-    ComboBoxMetrics(arrowMinSize, arrowSize, minSize, cornerSize, contentPadding, borderWidth, maxPopupHeight)
+): ComboBoxMetrics = ComboBoxMetrics(arrowAreaSize, minSize, cornerSize, contentPadding, borderWidth, maxPopupHeight)
 
 public fun ComboBoxMetrics.Companion.undecorated(
-    arrowMinSize: DpSize = DpSize(28.dp, 22.dp),
-    arrowSize: DpSize = DpSize(16.dp, 16.dp),
-    minSize: DpSize = DpSize((49 + 23 + 6).dp, 28.dp),
+    arrowAreaSize: DpSize = DpSize(23.dp, 24.dp),
+    minSize: DpSize = DpSize(49.dp, 24.dp),
     cornerSize: CornerSize = CornerSize(4.dp),
     contentPadding: PaddingValues = PaddingValues(horizontal = 6.dp, vertical = 2.dp),
     borderWidth: Dp = 0.dp,
     maxPopupHeight: Dp = 200.dp,
-): ComboBoxMetrics =
-    ComboBoxMetrics(arrowMinSize, arrowSize, minSize, cornerSize, contentPadding, borderWidth, maxPopupHeight)
+): ComboBoxMetrics = ComboBoxMetrics(arrowAreaSize, minSize, cornerSize, contentPadding, borderWidth, maxPopupHeight)
 
 public fun ComboBoxIcons.Companion.defaults(chevronDown: IconKey = AllIconsKeys.General.ChevronDown): ComboBoxIcons =
     ComboBoxIcons(chevronDown)
